refactor: move component reference signing into Core
The SignsComponentReferences contract lives in Core\Contracts and its
ComponentReferenceSigner implementation in Core\Services, dissolving the
single-class Security namespace.
