+++ src/Core/Concerns/InteractsWithLatticeComponents.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Bambamboole\Lattice\Core\Concerns;
 
 use Bambamboole\Lattice\Contracts\DefinitionRegistry;
-use Bambamboole\Lattice\Contracts\SignsComponentReferences;
+use Bambamboole\Lattice\Core\Contracts\SignsComponentReferences;
 use Bambamboole\Lattice\Core\Definition;
 use Bambamboole\Lattice\Exceptions\UnknownLatticeComponent;
 use Illuminate\Http\Request;
