+++ src/Http/Controllers/ActionController.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Bambamboole\Lattice\Http\Controllers;
 
 use Bambamboole\Lattice\Actions\ActionRegistry;
-use Bambamboole\Lattice\Contracts\SignsComponentReferences;
 use Bambamboole\Lattice\Core\Concerns\InteractsWithLatticeComponents;
+use Bambamboole\Lattice\Core\Contracts\SignsComponentReferences;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
