+++ src/LatticeServiceProvider.php
@@ -8,14 +8,14 @@ use BackedEnum;
 use Bambamboole\Lattice\Actions\ActionRegistry;
 use Bambamboole\Lattice\Actions\BulkActionRegistry;
 use Bambamboole\Lattice\Contracts\DiscoversDefinitions;
-use Bambamboole\Lattice\Contracts\SignsComponentReferences;
+use Bambamboole\Lattice\Core\Contracts\SignsComponentReferences;
+use Bambamboole\Lattice\Core\Services\ComponentReferenceSigner;
 use Bambamboole\Lattice\Discovery\DefinitionDiscovery;
 use Bambamboole\Lattice\Facades\Lattice;
 use Bambamboole\Lattice\Forms\FormRegistry;
 use Bambamboole\Lattice\Fragments\FragmentRegistry;
 use Bambamboole\Lattice\Menu\MenuItem;
 use Bambamboole\Lattice\Menu\MenuRegistry;
-use Bambamboole\Lattice\Security\ComponentReferenceSigner;
 use Bambamboole\Lattice\Tables\TableRegistry;
 use Closure;
 use Illuminate\Http\RedirectResponse;
