+++ tests/Feature/LatticePageTest.php
@@ -32,6 +32,7 @@ use Bambamboole\Lattice\Core\Enums\HttpMethod;
 use Bambamboole\Lattice\Core\Enums\LucideIcon;
 use Bambamboole\Lattice\Core\Enums\Width;
 use Bambamboole\Lattice\Core\PageSchema;
+use Bambamboole\Lattice\Core\Services\ComponentReferenceSigner;
 use Bambamboole\Lattice\Core\Toasts\Enums\ToastType;
 use Bambamboole\Lattice\Core\Toasts\ToastMessage;
 use Bambamboole\Lattice\Facades\Lattice;
@@ -45,7 +46,6 @@ use Bambamboole\Lattice\Http\Page;
 use Bambamboole\Lattice\LatticeRegistry;
 use Bambamboole\Lattice\Menu\MenuItem;
 use Bambamboole\Lattice\Menu\MenuRegistry;
-use Bambamboole\Lattice\Security\ComponentReferenceSigner;
 use Bambamboole\Lattice\Tables\Columns\StackColumn;
 use Bambamboole\Lattice\Tables\Columns\TextColumn;
 use Bambamboole\Lattice\Tables\Components\Table;
